fix: rename getDaysBetweenDates to calculateDaysDifferenceFromNow, relocate comparePurchaseUrgency to dates.jsx
